Add files via upload
Welcome to the Airbnb Clone repository built using HTML and CSS! This project aims to replicate the user interface and design elements of the popular Airbnb website, providing developers with a hands-on opportunity to enhance their HTML and CSS skills. Whether you're a beginner looking to practice your front-end development skills or an experienced developer exploring new projects, this Airbnb Clone is a great way to dive into web development.